fixed regression in decimal timescales, changing to long and add new Date(ts) to timescales caused exception for decimal timescales.  Added checkbox in configure to use literal text (without appending MorbidMeter) to the title.  Fixed widget to expand vertically if title overflows its line.

+++ morbidmeterlib/src/main/java/org/epstudios/morbidmeter/lib/CalendarTimeScale.java
@@ -41,13 +41,13 @@ public class CalendarTimeScale extends TimeScale {
 	}
 
 	@Override
-	public long proportionalTime(double percent) {
-		return minTime.getTimeInMillis() + (long)(percent * duration());
+	public double proportionalTime(double percent) {
+		return minTime.getTimeInMillis() + (percent * duration());
 	}
 
 	@Override
-	public long reverseProportionalTime(double percent) {
-		return maxTime.getTimeInMillis() - (long) (percent * duration());
+	public double reverseProportionalTime(double percent) {
+		return maxTime.getTimeInMillis() - (percent * duration());
 	}
 
 	private final Calendar minTime;
